requirements.txt
flask

langchain
langchain-community
langchain-core
langchainhub


langchain-groq
langchain-huggingface

dotenv


pandas

mysql-connector-python